Fix post-Queens server_group - server mismatch
After Queens upgrade, cloudmailer stopped working when launched with the
-y flag, which is when using it with a list of nodes in input.
Currently, cloudmailer assumes that a server that is a member of a
server group must be also in the "generic" list of servers, i.e., the
active non-deleted servers in the platform. This assumption is broken
after the Queens upgrade: the database has active server group whose
members correspond to deleted servers.

This patch takes into account the possiblity that a member of a server
group could not be in the list of servers retrieved from Openstack.
The servers of this kind can be simply discarded as they correspond to
deleted instances and thus not relevant for the cloudmailer use case.