Make email sending output more explicit
In case of error when sending emails out, currently it is not clear
which of the emails were sent out correctly and which not. No
information are print about sending in BCC.
This commit includes some status messages that help in understanding
where the error actually occurred and thus how to continue afterwards.