Fix emails being sent to spam if customer is internal
Emails having a proper DKIM signature do not get sent to spam. The DKIM
signature requires the "Subject" and the "From" fields to be declared in
the message, otherwise it cannot be done. Currently our messages have
only the Subject and the To fields declared, while the sender
information are only in the text field of the actual email. Due to this,
the emails do not get DKIM signature and get label as spam.
This commit adds the From field to the emails, thanks to which they get
a proper DKIM signature and do not get treated as spam any longer.